requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
torch==2.5.0+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.0+cu118
typing_extensions==4.9.0
tzdata==2024.2
